Move 'suffixesadd' from after/ftplugin to ftplugin
In #46, the question was where to put this option, and why some other
plugins put it in the 'after' directory. Vim-javascript puts it in the
'after' directory because they want to keep the settings in the ftplugin
shipped with Vim, and only add the 'suffixesadd' option. Since Vim
doesn't ship with a Vue plugin, we can safely put this option in ftplugin.