set "syntax sync fromstart" by default
Instead of setting it by default, I initially only added this as a tip
to the readme, because it is said to be slower. However, I don't really
notice a performance difference, and when the highlighting isn't working
reliably without it, I think it's better to just include it by default.