Add TSX support for <script type="tsx">
Previously, TSX wasn't being highlighted as when in typescript it requires the language to be `typescriptreact`, not `typescript`.

I've changed it so that the `typescript` matcher no longer matches `type="tsx"`, and there's a new matcher to handle that.

Honestly I have absolutely no idea what's going on in that regex so might have got something wrong. I've manually tested the expression against a few different script type types but don't 100% understand the intention of the original expression (which also matches `type="tshelloworld"`).